fix: change image source

+++ tools/scripts/src/update-latest.ts
@@ -1,22 +1,27 @@
 import { spawn } from 'child_process';
 
-const runCommand = (exec: string) => {
+const runCommand = (exec: string, interactive = false) => {
   const execParts = exec.split(' ');
   const [cmd, ...args] = execParts;
 
   return new Promise<void>((resolve, reject) => {
-    const child = spawn(cmd, args, { stdio: 'pipe', shell: true });
-
-    child.stdout.on('data', (data) => {
-      const text = data.toString();
-      process.stdout.write(text);
-      if (text.includes('STOP')) {
-        child.kill();
-        process.exit(0);
-      }
+    const child = spawn(cmd, args, {
+      stdio: interactive ? 'inherit' : 'pipe',
+      shell: true,
     });
 
-    child.stderr.on('data', (data) => process.stderr.write(data.toString()));
+    if (!interactive) {
+      child.stdout!.on('data', (data) => {
+        const text = data.toString();
+        process.stdout.write(text);
+        if (text.includes('STOP')) {
+          child.kill();
+          process.exit(0);
+        }
+      });
+
+      child.stderr!.on('data', (data) => process.stderr.write(data.toString()));
+    }
 
     child.on('close', (code) => {
       if (code !== 0) reject(new Error(`${cmd} exited with ${code}`));
@@ -31,5 +36,5 @@ const runCommand = (exec: string) => {
   await runCommand('pnpm pull-playlist -- section');
   await runCommand('pnpm pull-playlist -- character');
   await runCommand('pnpm get-img -- original');
-  await runCommand('pnpm set-series');
+  await runCommand('pnpm set-series', true); // 交互模式
 })();
